+++ src/editor/refresher/refresh_line.c
@@ -15,10 +15,10 @@
 
 static void	init_refresh(t_editor *e, t_refresher *r)
 {
-	r->plen = e->plen;
-	r->rows = (r->plen + e->len + e->cols - 1) / e->cols;
-	r->rpos = (r->plen + e->oldpos + e->cols) / e->cols;
-	r->old_rows = e->maxrows;
+	r->plen = (int)e->plen;
+	r->rows = (int)((r->plen + e->len + e->cols - 1) / e->cols);
+	r->rpos = (int)((r->plen + e->oldpos + e->cols) / e->cols);
+	r->old_rows = (int)e->maxrows;
 }
 
 static void	init_buf(t_buf *buf)
@@ -40,13 +40,13 @@ void		refresh_line(t_editor *e)
 	init_refresh(e, &ref);
 	init_buf(&buf);
 	if (ref.rows > (int)e->maxrows)
-		e->maxrows = ref.rows;
+		e->maxrows = (size_t)ref.rows;
 	clear_used_before(&ref, &buf);
 	clear_and_go_up(&ref, &buf);
 	clean_top_show_prompt(e, &buf);
 	insert_new_line(e, &ref, &buf);
 	move_cursor(e, &ref, &buf);
 	set_colum(e, &ref, &buf);
-	write(e->ofd, buf.b, buf.len);
+	write(e->ofd, buf.b, (size_t)buf.len);
 	free(buf.b);
 }
